refactor(notes): split notes-editor into window/tabs/content/footer
Phase 2 of frontend architecture refactor. Breaks the 910-line
notes-editor.tsx monolith into a feature folder:

- useNotesWindow / useNotesEditorInstance / useAiAssistant / useLinkEditor
  isolate window geometry, TipTap lifecycle, AI preview and link editing.
- NotesEditorTitleBar / Content / AiPreview / Footer / EditorBubbleMenu
  are standalone subcomponents under src/components/notes/NotesEditor/.
- NotesEditor.tsx is now a 218-line orchestrator wiring the four hooks
  to the five subcomponents, with only the delete-confirmation dialog
  and the 'close empty note' policy kept locally.

No behavior change. Progress tracked in docs/REFACTOR_FRONTEND_PROGRESS.md.
